From Rich Paul:
Sometimes routines allocate objects of type dep_t using xmalloc,
and not initializing all of the members:  the specific issue noticed
was that the ignore_mtime field, which marks a target as an 'order
only' target is not always initialized.  You can recognize that this is
happening by doing a make -p, and look for target lines like:

>> tgt:  dep1 | dep2

The dependency dep1 and it's ilk are normal deps, but changes to dep2 and
friends will not (if I understand order only targets properly) cause
tgt to be rebuilt ... their only contribution is that if, for some
reason, dep2 and tgt are rebuilt, the dep2 target will be completed
before tgt is started.

The problem is NOT limited to dep_t structures: if rather than
initializing all malloc'd memory to 0, I initialize it to 0xff, the
resulting make cannot even parse it's own src/Makefile: it fails with
a 'missing separator' message ... this appears to be a separate bug in
read.c, which has, as of yet, eluded me.